Add OVS cleanup utility
Fixes bug 1091605

The utility should be called after rebooting an appliance. This
will purge the openvswicth of configured tap devices.

A configuration variable quantum_ports has been added. This is
by default True which indicates that only Quantum ports will be
deleted from the OVS. If this is set as False then all ports on the
bridge will be deleted.

Change-Id: I442f64cf82f95bfa99d7765eb09db1ce2ecf602e